requirements.txt
fastapi
uvicorn
streamlit
gitpython
chromadb
sentence-transformers
langchain
groq
requests
python-dotenv